requirements.txt
matplotlib==3.4.3
numpy
paddlepaddle==2.1.3
pandas==1.3.3
Pillow==8.3.2
PyYAML==5.4.1
scikit-learn
scipy
tensorboardX
tqdm==4.62.3